# Conflicts:
#	.idea/workspace.xml
#	Domasna3/VinodventuraApp/src/main/java/mk/ukim/finki/vinodventuraapp/web/controller/RegisterController.java
#	Domasna3/VinodventuraApp/src/main/resources/templates/register.html

+++ Domasna3/VinodventuraApp/src/main/java/mk/ukim/finki/vinodventuraapp/web/controller/WishListController.java
@@ -20,7 +20,13 @@ public class WishListController {
     private final WishListService wishListService;
 
     @GetMapping
-    public String getWishList(Model model, @SessionAttribute User user) {
+    public String getWishList(Model model, @SessionAttribute(required = false) User user) {
+        if (user == null){
+            model.addAttribute("hasError", true);
+            model.addAttribute("error", "You need to be logged in to do this action.");
+            model.addAttribute("bodyContent", "home");
+            return "master-template";
+        }
         List<Winery> wineries = wishListService.getWishListForUser(user).getWineries();
         model.addAttribute("bodyContent", "wish-list");
         model.addAttribute("wineries", wineries);
@@ -29,19 +35,33 @@ public class WishListController {
     }
 
     @PostMapping("/add-winery/{id}")
-    public String addWineryToWishList(@PathVariable Long id, @SessionAttribute User user) {
-
+    public String addWineryToWishList(@PathVariable Long id, @SessionAttribute(required = false) User user,
+                                      Model model) {
+        if (user == null){
+            model.addAttribute("hasError", true);
+            model.addAttribute("error", "You need to be logged in to do this action.");
+            model.addAttribute("bodyContent", "home");
+            return "master-template";
+        }
         try {
             WishList wishList = this.wishListService.addToWishlist(user, id);
             return "redirect:/allWineries";
         } catch (RuntimeException exception) {
-
-            return "redirect:/allWineries?error=" + exception.getMessage();
+            model.addAttribute("hasError", true);
+            model.addAttribute("error", exception.getMessage());
+            model.addAttribute("bodyContent", "allWineries");
+            return "master-template";
         }
     }
 
     @PostMapping("/delete-winery/{id}")
     public String deleteProduct(@PathVariable Long id, @SessionAttribute User user, Model model) {
+        if (user == null){
+            model.addAttribute("hasError", true);
+            model.addAttribute("error", "You need to be logged in to do this action.");
+            model.addAttribute("bodyContent", "home");
+            return "master-template";
+        }
 
         this.wishListService.removeFromWishList(user, id);
